refactor(tags): service責務を分割 (#1307)
